Refactor to combine layout + scatter functions, mv rgb_columns out
