+++ osmo_jupyter/constants.py
@@ -1,0 +1,7 @@
+COLOR_CHANNELS = ['r', 'g', 'b']
+
+COLORS_BY_LETTER = {
+    'r': 'red',
+    'g': 'green',
+    'b': 'blue'
+}
